# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.UserCreateInput:
    properties:
      admin:
        type: boolean
      avatar:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      nickname:
        type: string
      password:
        type: string
      status:
        type: integer
      storage_quota:
        type: integer
      username:
        type: string
    type: object
  dto.UserDTO:
    properties:
      admin:
        type: boolean
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: integer
      nickname:
        type: string
      status:
        type: integer
      storage_quota:
        type: integer
      storage_used:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  dto.UserListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.UserDTO'
        type: array
      total:
        type: integer
    type: object
...
  dto.UserUpdateInput:
    properties:
      admin:
        type: boolean
      avatar:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      nickname:
        type: string
      password:
        type: string
      status:
        type: integer
      storage_quota:
        type: integer
      use_default_storage_quota:
        type: boolean
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 这是一个轻量级的说说服务
  title: EasyDrop API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      description: 分页查询用户列表，支持按用户名、邮箱、状态过滤
      parameters:
      - description: 用户名（模糊匹配）
        in: query
        name: username
        type: string
      - description: 邮箱（模糊匹配）
        in: query
        name: email
        type: string
      - description: 用户状态
        in: query
        name: status
        type: integer
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - description: 每页条数，最大 100
        in: query
        name: size
        type: integer
      - description: 排序，如 id desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserListResult'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端查询用户列表
      tags:
      - user-admin
    post:
      consumes:
      - application/json
      description: 创建用户并返回用户信息
      parameters:
      - description: 用户创建参数
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserDTO'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: 用户名或邮箱冲突
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端创建用户
      tags:
      - user-admin
  /admin/users/{id}:
    delete:
      description: 根据用户 ID 删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端删除用户
      tags:
      - user-admin
    patch:
      consumes:
      - application/json
      description: 根据用户 ID 更新用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户更新参数
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserDTO'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: 用户名或邮箱冲突
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端更新用户
      tags:
      - user-admin
  /admin/users/{id}/avatar:
    delete:
      description: 删除指定用户头像并更新存储占用
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端删除用户头像
      tags:
      - user-admin
    post:
      consumes:
      - multipart/form-data
      description: 为指定用户上传并替换头像
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 头像文件
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserDTO'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限或配额不足
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端上传用户头像
      tags:
      - user-admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"